+++ R/run_feature_filter.R
@@ -46,11 +46,10 @@ run_feature_filter <- function(qsip_data_object,
                                min_labeled_fractions = 2,
                                quiet = FALSE) {
 
-  # error ir is_qsip_data(qsip_data_object) is not true
+  # error if is_qsip_data(qsip_data_object) is not true
   if (isFALSE(is_qsip_data(qsip_data_object))) {
     stop("<qsip_data_object> must be of class qsip_data", call. = FALSE)
   }
-
 
   # make sure minimums are not bigger than possible
   if (min_labeled_sources > length(labeled_source_mat_ids)) {